# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Recipes API
  version: 1.0.0
  description: API para gestión de recetas, ingredientes, menús, usuarios, proveedores, inventario e impuestos.
servers:
  - url: http://localhost:4000
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
    InventoryMovement:
      type: object
      properties:
        movement_id:
          type: integer
        date:
          type: string
          format: date-time
        quantity:
          type: number
        movement_type:
          type: string
        comment:
          type: string
        recorded_by_user_id:
          type: integer
    InventoryMovementCreate:
      type: object
      required:
        - ingredient_id
        - date
        - quantity
        - movement_type
      properties:
        ingredient_id:
          type: integer
        date:
          type: string
          format: date-time
        quantity:
          type: number
        movement_type:
          type: string
        comment:
          type: string
    StockResponse:
      type: object
      properties:
        ingredient_id:
          type: integer
        stock:
          type: number
    Menu:
      type: object
      properties:
        menu_id:
          type: integer
        name:
          type: string
        menu_date:
          type: string
          format: date
    MenuCreate:
      type: object
      required:
        - name
        - menu_date
      properties:
        name:
          type: string
        menu_date:
          type: string
          format: date
    MenuRecipeItem:
      type: object
      properties:
        recipe_id:
          type: integer
        name:
          type: string
        net_price:
          type: number
        cost_per_serving:
          type: number
...
paths:
...
  /inventory/{ingredient_id}:
    get:
      tags:
        - Inventario
      summary: Historial de movimientos
      description: |
        Recupera el historial de movimientos de inventario de un ingrediente.
        **Roles permitidos**: `admin`, `inventory_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryMovement'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/stock/{ingredient_id}:
    get:
      tags:
        - Inventario
      summary: Stock actual
      description: |
        Obtiene el stock actual calculado para un ingrediente sumando entradas y ajustes y restando salidas.
        **Roles permitidos**: `admin`, `inventory_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stock actual del ingrediente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory:
    post:
      tags:
        - Inventario
      summary: Registrar nuevo movimiento
      description: |
        Registra un nuevo movimiento de inventario (entrada, ajuste o salida).
        **Roles permitidos**: `admin`, `inventory_manager`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryMovementCreate'
      responses:
        '201':
          description: Movimiento registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menus:
    get:
      tags:
        - Menús
      summary: Listar menús
      description: |
        Recupera todos los menús ordenados por fecha.
        **Roles permitidos**: `admin`, `chef`, `waiter`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de menús
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'

    post:
      tags:
        - Menús
      summary: Crear menú
      description: |
        Crea un nuevo menú con nombre y fecha.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCreate'
      responses:
        '201':
          description: Menú creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menus/{id}:
    get:
      tags:
        - Menús
      summary: Obtener menú por ID
      description: |
        Recupera los datos de un menú específico.
        **Roles permitidos**: `admin`, `chef`, `waiter`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del menú
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          description: Menú no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Menús
      summary: Actualizar menú
      description: |
        Modifica nombre y/o fecha de un menú existente.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCreate'
      responses:
        '200':
          description: Menú actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Menú no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Menús
      summary: Eliminar menú
      description: |
        Elimina un menú y sus asociaciones de recetas.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Menú eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Menú no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menus/{id}/recipes:
    get:
      tags:
        - Menús
      summary: Listar recetas de un menú
      description: |
        Obtiene las recetas asignadas a un menú.
        **Roles permitidos**: `admin`, `chef`, `waiter`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Recetas del menú
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuRecipeItem'

    post:
      tags:
        - Menús
      summary: Asignar receta a menú
      description: |
        Añade una receta existente a un menú.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipe_id
              properties:
                recipe_id:
                  type: integer
      responses:
        '201':
          description: Receta asignada al menú
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: La receta ya está asignada o parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /menus/{id}/recipes/{recipe_id}:
    delete:
      tags:
        - Menús
      summary: Quitar receta de menú
      description: |
        Elimina la asignación de una receta a un menú.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: recipe_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Receta eliminada del menú
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Menú o receta no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...